* KDE/4.14:
  kio_sftp: Use the right type for timeout_sec and timeout_usec.
  SVN_SILENT made messages (.desktop file)

+++ kioslave/sftp/kio_sftp.cpp
@@ -528,7 +528,7 @@ bool sftpProtocol::sftpOpenConnection (const AuthInfo& info)
     return false;
   }
 
-  int timeout_sec = 30, timeout_usec = 0;
+  long timeout_sec = 30, timeout_usec = 0;
 
   kDebug(KIO_SFTP_DB) << "Creating the SSH session and setting options";
